+++ src/Schema.ts
@@ -41,7 +41,7 @@ import * as InternalSchema from "./internal/schema.js"
 import * as InternalSerializable from "./internal/serializable.js"
 import * as Parser from "./Parser.js"
 import * as ParseResult from "./ParseResult.js"
-import type { Pretty } from "./Pretty.js"
+import * as Pretty from "./Pretty.js"
 import type * as Serializable from "./Serializable.js"
 
 // ---------------------------------------------
@@ -720,7 +720,7 @@ export class PropertySignatureImpl<From, FromIsOptional, To, ToIsOptional> {
  */
 export const propertySignature = <I, A>(
   schema: Schema<I, A>,
-  options: DocAnnotations<A>
+  options: DocAnnotations
 ): PropertySignature<I, false, A, false> =>
   new PropertySignatureImpl({
     _tag: "PropertySignature",
@@ -733,7 +733,7 @@ export const propertySignature = <I, A>(
  */
 export const optional = <I, A>(
   schema: Schema<I, A>,
-  options?: DocAnnotations<A>
+  options?: DocAnnotations
 ): OptionalPropertySignature<I, true, A, true> =>
   new PropertySignatureImpl({
     _tag: "Optional",
@@ -940,10 +940,10 @@ export interface BrandSchema<From, To extends Brand.Brand<any>>
   extends Schema<From, To>, Brand.Brand.Constructor<To>
 {}
 
-const appendBrandAnnotation = <B extends string | symbol, A>(
+const appendBrandAnnotation = <B extends string | symbol>(
   ast: AST.AST,
   brand: B,
-  options?: DocAnnotations<A>
+  options?: DocAnnotations
 ): AST.AST => {
   if (AST.isTransform(ast)) {
     return AST.createTransform(
@@ -979,7 +979,7 @@ const appendBrandAnnotation = <B extends string | symbol, A>(
  */
 export const brand = <B extends string | symbol, A>(
   brand: B,
-  options?: DocAnnotations<A>
+  options?: DocAnnotations
 ) =>
 <I>(self: Schema<I, A>): BrandSchema<I, A & Brand.Brand<B>> => {
   const ast = appendBrandAnnotation(self.ast, brand, options)
@@ -1482,20 +1482,20 @@ const toAnnotations = <A>(
 /**
  * @since 1.0.0
  */
-export interface DocAnnotations<A> extends AST.Annotations {
+export interface DocAnnotations extends AST.Annotations {
   readonly identifier?: AST.IdentifierAnnotation
   readonly title?: AST.TitleAnnotation
   readonly description?: AST.DescriptionAnnotation
   readonly examples?: AST.ExamplesAnnotation
   readonly default?: AST.DefaultAnnotation
   readonly documentation?: AST.DocumentationAnnotation
-  readonly message?: AST.MessageAnnotation<A>
 }
 
 /**
  * @since 1.0.0
  */
-export interface FilterAnnotations<A> extends DocAnnotations<A> {
+export interface FilterAnnotations<A> extends DocAnnotations {
+  readonly message?: AST.MessageAnnotation<A>
   readonly typeId?: AST.TypeAnnotation | { id: AST.TypeAnnotation; params: unknown }
   /**
    * Attaches a JSON Schema annotation to this refinement.
@@ -1504,7 +1504,7 @@ export interface FilterAnnotations<A> extends DocAnnotations<A> {
    */
   readonly jsonSchema?: AST.JSONSchemaAnnotation
   readonly arbitrary?: (...args: ReadonlyArray<Arbitrary<any>>) => Arbitrary<any>
-  readonly pretty?: (...args: ReadonlyArray<Pretty<any>>) => Pretty<any>
+  readonly pretty?: (...args: ReadonlyArray<Pretty.Pretty<any>>) => Pretty.Pretty<any>
   readonly equivalence?: () => Equivalence.Equivalence<A>
 }
 
@@ -3011,7 +3011,7 @@ export const SecretFromSelf: Schema<Secret.Secret> = declare(
       : ParseResult.fail(ParseResult.type(ast, u)),
   {
     [AST.IdentifierAnnotationId]: "Secret",
-    [hooks.PrettyHookId]: (): Pretty<Secret.Secret> => (secret) => String(secret),
+    [hooks.PrettyHookId]: (): Pretty.Pretty<Secret.Secret> => (secret) => String(secret),
     [hooks.ArbitraryHookId]: (): Arbitrary<Secret.Secret> => (fc) =>
       fc.string().map((_) => Secret.fromString(_))
   }
@@ -3063,7 +3063,7 @@ export const DurationFromSelf: Schema<Duration.Duration> = declare(
       : ParseResult.fail(ParseResult.type(ast, u)),
   {
     [AST.IdentifierAnnotationId]: "Duration",
-    [hooks.PrettyHookId]: (): Pretty<Duration.Duration> => (duration) => String(duration),
+    [hooks.PrettyHookId]: (): Pretty.Pretty<Duration.Duration> => (duration) => String(duration),
     [hooks.ArbitraryHookId]: (): Arbitrary<Duration.Duration> => (fc) =>
       fc.oneof(
         fc.constant(Duration.infinity),
@@ -3356,7 +3356,7 @@ export const Uint8ArrayFromSelf: Schema<Uint8Array> = declare(
       : ParseResult.fail(ParseResult.type(ast, u)),
   {
     [AST.IdentifierAnnotationId]: "Uint8Array",
-    [hooks.PrettyHookId]: (): Pretty<Uint8Array> => (u8arr) =>
+    [hooks.PrettyHookId]: (): Pretty.Pretty<Uint8Array> => (u8arr) =>
       `new Uint8Array(${JSON.stringify(Array.from(u8arr))})`,
     [hooks.ArbitraryHookId]: (): Arbitrary<Uint8Array> => (fc) => fc.uint8Array(),
     [hooks.EquivalenceHookId]: () => ReadonlyArray.getEquivalence(Equivalence.strict())
@@ -3429,7 +3429,7 @@ const makeEncodingTransform = <A extends string>(
     { strict: false }
   ).pipe(annotations({
     [AST.IdentifierAnnotationId]: id,
-    [hooks.PrettyHookId]: (): Pretty<Uint8Array> => (u) => `${id}(${encode(u)})`,
+    [hooks.PrettyHookId]: (): Pretty.Pretty<Uint8Array> => (u) => `${id}(${encode(u)})`,
     [hooks.ArbitraryHookId]: () => arbitrary
   }))
 
@@ -3623,7 +3623,7 @@ export const validDate =
 
 const dateArbitrary = (): Arbitrary<Date> => (fc) => fc.date({ noInvalidDate: false })
 
-const datePretty = (): Pretty<Date> => (date) => `new Date(${JSON.stringify(date)})`
+const datePretty = (): Pretty.Pretty<Date> => (date) => `new Date(${JSON.stringify(date)})`
 
 /**
  * Represents a schema for handling potentially **invalid** `Date` instances (e.g., `new Date("Invalid Date")` is not rejected).
@@ -3729,7 +3729,7 @@ const optionArbitrary = <A>(value: Arbitrary<A>): Arbitrary<Option.Option<A>> =>
   return (fc) => arb(fc).map(optionDecode)
 }
 
-const optionPretty = <A>(value: Pretty<A>): Pretty<Option.Option<A>> =>
+const optionPretty = <A>(value: Pretty.Pretty<A>): Pretty.Pretty<Option.Option<A>> =>
   Option.match({
     onNone: () => "none()",
     onSome: (a) => `some(${value(a)})`
@@ -3833,7 +3833,10 @@ const eitherArbitrary = <E, A>(
   return (fc) => arb(fc).map(eitherDecode)
 }
 
-const eitherPretty = <E, A>(left: Pretty<E>, right: Pretty<A>): Pretty<Either.Either<E, A>> =>
+const eitherPretty = <E, A>(
+  left: Pretty.Pretty<E>,
+  right: Pretty.Pretty<A>
+): Pretty.Pretty<Either.Either<E, A>> =>
   Either.match({
     onLeft: (e) => `left(${left(e)})`,
     onRight: (a) => `right(${right(a)})`
@@ -3900,9 +3903,9 @@ const readonlyMapArbitrary = <K, V>(
 (fc) => fc.array(fc.tuple(key(fc), value(fc))).map((as) => new Map(as))
 
 const readonlyMapPretty = <K, V>(
-  key: Pretty<K>,
-  value: Pretty<V>
-): Pretty<ReadonlyMap<K, V>> =>
+  key: Pretty.Pretty<K>,
+  value: Pretty.Pretty<V>
+): Pretty.Pretty<ReadonlyMap<K, V>> =>
 (map) =>
   `new Map([${
     Array.from(map.entries())
@@ -3975,7 +3978,7 @@ const isSet = (u: unknown): u is Set<unknown> => u instanceof Set
 const readonlySetArbitrary = <A>(item: Arbitrary<A>): Arbitrary<ReadonlySet<A>> => (fc) =>
   fc.array(item(fc)).map((as) => new Set(as))
 
-const readonlySetPretty = <A>(item: Pretty<A>): Pretty<ReadonlySet<A>> => (set) =>
+const readonlySetPretty = <A>(item: Pretty.Pretty<A>): Pretty.Pretty<ReadonlySet<A>> => (set) =>
   `new Set([${Array.from(set.values()).map((a) => item(a)).join(", ")}])`
 
 const readonlySetEquivalence = <A>(
@@ -4032,7 +4035,7 @@ export const readonlySet = <I, A>(item: Schema<I, A>): Schema<ReadonlyArray<I>, 
 // BigDecimal transformations
 // ---------------------------------------------
 
-const bigDecimalPretty = (): Pretty<BigDecimal.BigDecimal> => (val) =>
+const bigDecimalPretty = (): Pretty.Pretty<BigDecimal.BigDecimal> => (val) =>
   `BigDecimal(${BigDecimal.format(BigDecimal.normalize(val))})`
 
 const bigDecimalArbitrary = (): Arbitrary<BigDecimal.BigDecimal> => (fc) =>
@@ -4376,7 +4379,7 @@ export const negateBigDecimal = <I, A extends BigDecimal.BigDecimal>(
 const chunkArbitrary = <A>(item: Arbitrary<A>): Arbitrary<Chunk.Chunk<A>> => (fc) =>
   fc.array(item(fc)).map(Chunk.fromIterable)
 
-const chunkPretty = <A>(item: Pretty<A>): Pretty<Chunk.Chunk<A>> => (c) =>
+const chunkPretty = <A>(item: Pretty.Pretty<A>): Pretty.Pretty<Chunk.Chunk<A>> => (c) =>
   `Chunk(${Chunk.toReadonlyArray(c).map(item).join(", ")})`
 
 /**
@@ -4435,8 +4438,8 @@ const dataArbitrary = <A extends Readonly<Record<string, any>> | ReadonlyArray<a
 (fc) => item(fc).map(toData)
 
 const dataPretty = <A extends Readonly<Record<string, any>> | ReadonlyArray<any>>(
-  item: Pretty<A>
-): Pretty<Data.Data<A>> =>
+  item: Pretty.Pretty<A>
+): Pretty.Pretty<Data.Data<A>> =>
 (d) => `Data(${item(d)})`
 
 /**
@@ -4635,7 +4638,9 @@ export declare namespace TaggedRequest {
    * @category classes
    * @since 1.0.0
    */
-  export type Any = TaggedRequest<string, any, any, any, any, any, any>
+  export type Any =
+    | TaggedRequest<string, any, any, any, any, any, any>
+    | TaggedRequest<string, any, any, never, never, any, any>
 }
 
 /**
@@ -4702,6 +4707,10 @@ const makeClass = <I, A>(
 
     static [TypeId] = variance
 
+    toString() {
+      return Pretty.to(this.constructor as any)(this)
+    }
+
     static pipe() {
       return pipeArguments(this, arguments)
     }
@@ -4715,6 +4724,8 @@ const makeClass = <I, A>(
             ParseResult.succeed(input)
             : ParseResult.fail(ParseResult.type(ast, input)), {
           [AST.DescriptionAnnotationId]: `an instance of ${this.name}`,
+          [hooks.PrettyHookId]: (struct: any) => (self: any) =>
+            `${self.constructor.name}(${struct(self)})`,
           [hooks.ArbitraryHookId]: (struct: any) => (fc: any) =>
             struct(fc).map((props: any) => new this(props))
         }),
@@ -4823,7 +4834,7 @@ const fiberIdFromArbitrary = arbitrary.unsafe(FiberIdFrom)
 const fiberIdArbitrary: Arbitrary<FiberId.FiberId> = (fc) =>
   fiberIdFromArbitrary(fc).map(fiberIdDecode)
 
-const fiberIdPretty: Pretty<FiberId.FiberId> = (fiberId) => {
+const fiberIdPretty: Pretty.Pretty<FiberId.FiberId> = (fiberId) => {
   switch (fiberId._tag) {
     case "None":
       return "FiberId.none"
@@ -4971,7 +4982,7 @@ const causeArbitrary = <E>(
   return (fc) => arb(fc).map(causeDecode)
 }
 
-const causePretty = <E>(error: Pretty<E>): Pretty<Cause.Cause<E>> => (cause) => {
+const causePretty = <E>(error: Pretty.Pretty<E>): Pretty.Pretty<Cause.Cause<E>> => (cause) => {
   const f = (cause: Cause.Cause<E>): string => {
     switch (cause._tag) {
       case "Empty":
@@ -5143,10 +5154,12 @@ const exitArbitrary = <E, A>(
   return (fc) => arb(fc).map(exitDecode)
 }
 
-const exitPretty = <E, A>(error: Pretty<E>, value: Pretty<A>): Pretty<Exit.Exit<E, A>> => (exit) =>
-  exit._tag === "Failure"
-    ? `Exit.failCause(${causePretty(error)(exit.cause)})`
-    : `Exit.succeed(${value(exit.value)})`
+const exitPretty =
+  <E, A>(error: Pretty.Pretty<E>, value: Pretty.Pretty<A>): Pretty.Pretty<Exit.Exit<E, A>> =>
+  (exit) =>
+    exit._tag === "Failure"
+      ? `Exit.failCause(${causePretty(error)(exit.cause)})`
+      : `Exit.succeed(${value(exit.value)})`
 
 /**
  * @category Exit